Scripts/requirements.txt
google-genai
pathlib